Rename util.js -> util.js
